pyproject.toml
[project]
name = "test-101"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.2.14",
    "langgraph>=0.6.7",
    "neatlogs>=1.1.7",
    "python-dotenv>=1.1.1",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "schedule>=1.2.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "openpyxl>=3.1.0",
]
